+++ zeyvrometacounter.php
@@ -27,7 +27,7 @@ class Zeyvrometacounter extends Module
     {
         $this->name = 'zeyvrometacounter';
         $this->tab = 'seo';
-        $this->version = '1.0.5';
+        $this->version = '1.0.6';
         $this->author = 'Zeyvro';
         $this->need_instance = 0;
         $this->ps_versions_compliancy = ['min' => '8.0.0', 'max' => '9.99.99'];